+++ appEditors/AppGerberEditor.py
@@ -3323,9 +3323,9 @@ class AppGerberEditor(QtCore.QObject):
         self.apertures_table.clearSelection()
 
         # Remove anything else in the GUI Properties Tab
-        self.app.ui.selected_scroll_area.takeWidget()
+        self.app.ui.properties_scroll_area.takeWidget()
         # Put ourselves in the GUI Properties Tab
-        self.app.ui.selected_scroll_area.setWidget(self.grb_edit_widget)
+        self.app.ui.properties_scroll_area.setWidget(self.grb_edit_widget)
         # Switch notebook to Properties page
         self.app.ui.notebook.setCurrentWidget(self.app.ui.properties_tab)